sendto to functional component
