+++ app/components/Views/SendFlow/AddressTo/types.ts
@@ -1,3 +1,12 @@
 export interface SFAddressToProps {
-    
+    inputRef: any;
+     highlighted: boolean;
+     addressToReady: boolean;
+     toSelectedAddress: string;
+     toAddressName: string;
+     onSubmit: (address: string) => void;
+     inputWidth: any;
+     confusableCollection: [];
+     isFromAddressBook: any;
+     updateParentState: any;
 }
